+++ package.json
@@ -1,22 +1,45 @@
 {
   "name": "@adefee/palworld-save-editor",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Comprehensive and extensible save editor and reporting tool for Palworld",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "type": "commonjs",
   "bin": "./index.js",
+  "exports": {
+    ".": {
+      "require": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
   "scripts": {
+    "build": "npx tsc",
+    "dev": "concurrently \"npx tsc --watch\" \"nodemon\"",
     "install-packager": "npm install -g pkg",
-    "package": "pkg . --targets node18-win-x64 --output ./bin/paver-win --compress Brotli",
-    "package-all": "pkg . --targets node18-linux-x64,node18-macos-x64,node18-win-x64 --output ./bin/paver --compress Brotli",
+    "prepackage": "cd dist && yarn install && cd ..",
+    "package": "pkg ./dist --targets node18-win-x64 --output ./bin/paver-win --compress Brotli",
+    "package-all": "pkg ./dist --targets node18-linux-x64,node18-macos-x64,node18-win-x64 --output ./bin/paver --compress Brotli",
     "start": "node .",
+    "start:s": "npx tsc && node ./dist/indexStreaming.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "GPL-3.0",
   "dependencies": {
-    "dotenv": "^16.4.1",
-    "minimist": "^1.2.8",
-    "readline": "^1.3.0"
-  }
+    "JSONStream": "^1.3.5",
+    "adm-zip": "^0.5.10",
+    "axios": "^1.6.7"
+  },
+  "devDependencies": {
+    "@types/node": "^20.11.16",
+    "@types/stream-chain": "^2.0.4",
+    "@types/stream-json": "^1.7.7",
+    "concurrently": "^8.2.2",
+    "nodemon": "^3.0.3",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
+  },
+  "types": "dist/index.d.ts"
 }
