+++ src/lib/filenameCompatibleTimestamp.js
@@ -1,0 +1,13 @@
+const filenameCompatibleTimestamp = () => {
+  const now = new Date();
+  const year = now.getFullYear();
+  const month = String(now.getMonth() + 1).padStart(2, '0'); // months are zero-based in JavaScript
+  const day = String(now.getDate()).padStart(2, '0');
+  const hours = String(now.getHours()).padStart(2, '0');
+  const minutes = String(now.getMinutes()).padStart(2, '0');
+  const seconds = String(now.getSeconds()).padStart(2, '0');
+
+  return `pal-${year}-${month}-${day}_${hours}${minutes}${seconds}`;
+}
+
+module.exports = filenameCompatibleTimestamp;
