+++ tsconfig.json
@@ -1,0 +1,16 @@
+{
+    "compilerOptions": {
+        "rootDir": "./src/",
+        "outDir": "dist",
+        "allowJs": true,
+        "moduleResolution": "node",
+        "allowSyntheticDefaultImports": true,
+        "esModuleInterop": true,
+        "sourceMap": true,
+        "module": "commonjs",
+        "target": "es2015",
+        "resolveJsonModule": true,
+        "declaration": true
+    },
+    "include": ["./src/**/*"]
+}
